BUG,ENH: Allow combination of double and unit for many ufuncs (and fix last mess)
The last mess was that I forgot to dereference in one place :(.
This adds a simple "homogeneous" promoter.  This is a bit of a hack,
not because it is wrong, but because I insert it to be applied to
all inputs.
NumPy should do this, not us here.  And NumPy *will* do this in the
future.
(If we needed a promoter here, it would have to be more specific than
this, which is possible, although slightly more annoying :))